Purge Compass from Gulp build! Begone, ye daemon!
It has been replaced by ruby-sass and autoprefixer. At this time, there
is a slight bug in ruby-sass, meaning it doesn’t allow autoprefixer and
source maps to play well together.

If this is not fixed, I will contemplate replacing gulp-ruby-sass  with
gulp-sass and gulp-sourcemaps.

gulp-ruby-sass was chosen, simply because it looks a bit neater, but I
can go either way.